+++ src/client/pages/CycleLanding/KeyFindings/keyFindings.scss
